Fixed recv functions

+++ socked.h
@@ -89,10 +89,10 @@ void skdSend(SkdSocket& skt, const char* msg, size_t size, int flags);
 void skdSendTo(SkdSocket& skt, const char* msg, size_t size, int flags);
 
 // Receive data from TCP connection
-uint64_t skdReceive(SkdSocket& skt, char* buffer, int size, int flags);
+int skdReceive(SkdSocket& skt, char* buffer, size_t size, int flags);
 
 // Receive data from Datagram connection.
-uint64_t skdReceiveFrom(SkdSocket& skt, char* buffer, size_t size, int flags);
+int skdReceiveFrom(SkdSocket& skt, char* buffer, size_t size, int flags);
 
 #endif // !SOCKED_H
